added status service
