+++ src/app/admin/services/core/status.service.spec.ts
@@ -1,0 +1,12 @@
+import { TestBed } from '@angular/core/testing';
+
+import { StatusService } from './status.service';
+
+describe('StatusService', () => {
+  beforeEach(() => TestBed.configureTestingModule({}));
+
+  it('should be created', () => {
+    const service: StatusService = TestBed.get(StatusService);
+    expect(service).toBeTruthy();
+  });
+});
